+++ lua/lazyvim/plugins/extras/editor/outline.lua
@@ -14,28 +14,21 @@ return {
     opts = function()
       local defaults = require("outline.config").defaults
       local opts = {
-        symbols = {},
-        symbol_blacklist = {},
+        symbols = {
+          icons = {},
+          filter = vim.deepcopy(LazyVim.config.kind_filter),
+        },
         keymaps = {
           up_and_jump = "<up>",
           down_and_jump = "<down>",
         },
       }
-      local filter = LazyVim.config.kind_filter
 
-      if type(filter) == "table" then
-        filter = filter.default
-        if type(filter) == "table" then
-          for kind, symbol in pairs(defaults.symbols) do
-            opts.symbols[kind] = {
-              icon = LazyVim.config.icons.kinds[kind] or symbol.icon,
-              hl = symbol.hl,
-            }
-            if not vim.tbl_contains(filter, kind) then
-              table.insert(opts.symbol_blacklist, kind)
-            end
-          end
-        end
+      for kind, symbol in pairs(defaults.symbols.icons) do
+        opts.symbols.icons[kind] = {
+          icon = LazyVim.config.icons.kinds[kind] or symbol.icon,
+          hl = symbol.hl,
+        }
       end
       return opts
     end,