+++ lua/lazyvim/util/lazygit.lua
@@ -154,9 +154,32 @@ function M.blame_line(opts)
   local cursor = vim.api.nvim_win_get_cursor(0)
   local line = cursor[1]
   local file = vim.api.nvim_buf_get_name(0)
-  local root = LazyVim.root.detectors.pattern(0, { ".git" })[1]
+  local root = LazyVim.root.detectors.pattern(0, { ".git" })[1] or "."
   local cmd = { "git", "-C", root, "log", "-n", opts.count, "-u", "-L", line .. ",+1:" .. file }
   return require("lazy.util").float_cmd(cmd, opts)
+end
+
+-- stylua: ignore
+M.remote_patterns = {
+  { "^(https?://.*)%.git$"              , "%1" },
+  { "^git@(.+):(.+)%.git$"              , "https://%1/%2" },
+  { "^git@(.+):(.+)$"                   , "https://%1/%2" },
+  { "^git@(.+)/(.+)$"                   , "https://%1/%2" },
+  { "^ssh://git@(.*)$"                  , "https://%1" },
+  { "ssh%.dev%.azure%.com/v3/(.*)/(.*)$", "dev.azure.com/%1/_git/%2" },
+  { "^https://%w*@(.*)"                 , "https://%1" },
+  { "^git@(.*)"                         , "https://%1" },
+  { ":%d+"                              , "" },
+  { "%.git$"                            , "" },
+}
+
+---@param remote string
+function M.get_url(remote)
+  local ret = remote
+  for _, pattern in ipairs(M.remote_patterns) do
+    ret = ret:gsub(pattern[1], pattern[2])
+  end
+  return ret:find("https://") == 1 and ret or ("https://%s"):format(ret)
 end
 
 function M.browse()
@@ -164,12 +187,15 @@ function M.browse()
   local remotes = {} ---@type {name:string, url:string}[]
 
   for _, line in ipairs(lines) do
-    local name, url = line:match("(%S+)%s+(%S+)%s+%(fetch%)")
-    if name and url then
-      if url:find("git@github.com") or url:find("git@bitbucket.org") or url:find("git@gitlab.com") then
-        url = url:gsub("git@(%S+):", "https://%1/"):gsub(".git$", "")
+    local name, remote = line:match("(%S+)%s+(%S+)%s+%(fetch%)")
+    if name and remote then
+      local url = M.get_url(remote)
+      if url then
+        table.insert(remotes, {
+          name = name,
+          url = url,
+        })
       end
-      table.insert(remotes, { name = name, url = url })
     end
   end